Move fixtures to tests folder
