+++ src/containers/pages/IRSLite/MapLite/tests/index.test.tsx
@@ -33,7 +33,7 @@ import jurisdictionReducer, {
   reducerName as jurisdictionReducerName,
 } from '../../../../../store/ducks/jurisdictions';
 import { InterventionType } from '../../../../../store/ducks/plans';
-import * as fixtures from '../../JurisdictionsReportLite/fixtures';
+import * as fixtures from '../../JurisdictionsReportLite/tests/fixtures';
 
 /* tslint:disable-next-line no-var-requires */
 const fetch = require('jest-fetch-mock');
